% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_convert.R
\name{id_convert}
\alias{id_convert}
\title{id conversion}
\usage{
id_convert(genelist = NULL, fromType = "ensembl_gene_id",
  species = "hs", alias = FALSE)
}
\arguments{
\item{genelist}{a vector containing genes you want to convert from}

\item{fromType}{\emph{genelist} type, one of ensemble_gene_id (default), symbol, entrezgene}

\item{species}{only recept hs (default), mm}

\item{alias}{if \emph{genelist} contains gene symbol alias (defalut is TRUE)}
}
\description{
convert given IDs (including alias) to specific IDs
}
