% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmid_construct.R
\name{primer_design}
\alias{primer_design}
\title{construct plasmid}
\usage{
primer_design(splashRNA, control_fa)
}
\arguments{
\item{splashRNA}{splashRNA_batch result}

\item{vector}{vector fasta}
}
\description{
construct a DNA fragment to a vector
}
