% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shRNA_primer.R
\name{shRNA_primer}
\alias{shRNA_primer}
\title{design primer}
\usage{
shRNA_primer(splashRNA, vector)
}
\arguments{
\item{splashRNA}{splashRNA result}

\item{vector}{vector fasta sequence}
}
\description{
design primer according to splashRNA result
}
