% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashRNA.R
\name{splashRNA}
\alias{splashRNA}
\title{search splashRNA for a single query}
\usage{
splashRNA(id, n, anno)
}
\arguments{
\item{id}{entrezgene id}

\item{n}{number of returned shRNA}

\item{anno}{names for each shRNA}
}
\description{
search splashRNA for a single query
}
