% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashRNA.R
\name{splashRNA_batch}
\alias{splashRNA_batch}
\title{search splashRNA for batch query}
\usage{
splashRNA_batch(ID, N, Anno)
}
\arguments{
\item{ID}{entrezgene ids}

\item{N}{number of returned shRNA}

\item{Anno}{names for each shRNA}
}
\description{
search splashRNA for batch query
}
